Finish arms journey

+++ app/routes/apply.js
@@ -597,27 +597,32 @@ router.get('/prove-eligibility/delete-medication/:id', function(req, res) {
 router.get('/prove-eligibility/list-healthcare-professionals', function(req, res) {
   var hcps = req.session.data['hcp-array'];
   delete res.locals.tableRows;
-  if (hcps) {
-    var tableRows = [];
-    hcps.forEach(function(item,index) {
-    tableRows.push([
-        {
-          "text": item.name
-        },
-        {
-          "text": item.hospital
-        },
-        {
-          "html": "<a href='delete-hcp/"+index+"'>Remove this</a>",
-          "format": "numeric"
-        }
-      ])
-    });
-  }
 
-  res.locals.tableRows = tableRows;
-  res.locals.formAction = '/apply-for-a-blue-badge/prepare/check-answers'
-  res.render(proveEligibilityTemplatePath+'list-healthcare-professionals');
+  if (req.session.data['treatments-array'] || req.session.data['medication-array']) {
+    if (hcps) {
+      var tableRows = [];
+      hcps.forEach(function(item,index) {
+      tableRows.push([
+          {
+            "text": item.name
+          },
+          {
+            "text": item.hospital
+          },
+          {
+            "html": "<a href='delete-hcp/"+index+"'>Remove this</a>",
+            "format": "numeric"
+          }
+        ])
+      });
+    }
+
+    res.locals.tableRows = tableRows;
+    res.locals.formAction = '/apply-for-a-blue-badge/prepare/check-answers'
+    res.render(proveEligibilityTemplatePath+'list-healthcare-professionals');
+  } else {
+    res.redirect('/apply-for-a-blue-badge/prepare/check-answers');
+  }
 });
 
 router.get('/prove-eligibility/add-healthcare-professional', function(req, res) {
